Use same port of video content ingestion for the telemetry HTTP server (#31)
AWS Load balancers in UDP mode expects the service to respond to health checks using TCP in the same port where the UDP service is exposed.

This is causing an issue where the load balancers created by the Hawkeye service are marked as unhealthy. Sometimes the load balancer will not forward the content to the target group. It is not clear why some load balancers still forward content when the target group is unhealthy. I suspect it can forward if content is sent right after the load balancer is created. But if the load balancer is created and several minutes later we start sending content, the target groups are already marked as unhealthy, the load balancer will not forward the video content.

This PR makes the Hawkeye Worker use the same ingest port for the HTTP telemetry service that expose metrics and the latest video frame captured. Since the GStreamer pipeline and the HTTP server are running in the same process we have no problem in binding both sockets to the same port. The content is not mixed since they use different network protocols.

We expect that after this PR the target groups of the AWS Load balancers get marked as healthy and the content to be forward without problems to the Hawkeye Workers.